create noteService, note dto and dao

+++ user-service/src/main/java/com/andzwp/userservice/dto/Role.java
@@ -3,5 +3,5 @@ package com.andzwp.userservice.dto;
 import lombok.Builder;
 
 @Builder
-public record Role(long id,String name) {
+public record Role(int id,String name) {
 }
